app/requirements.txt
fastapi
uvicorn
requests
PyYAML